# Conflicts:
#	ReactAndroid/src/main/java/com/facebook/react/views/text/ReactTextViewManager.java

+++ ReactAndroid/src/main/java/com/facebook/react/views/textinput/ReactContentSizeChangedEvent.java
@@ -21,13 +21,13 @@ public class ReactContentSizeChangedEvent extends Event<ReactTextChangedEvent> {
 
   public static final String EVENT_NAME = "topContentSizeChange";
 
-  private int mContentWidth;
-  private int mContentHeight;
+  private float mContentWidth;
+  private float mContentHeight;
 
   public ReactContentSizeChangedEvent(
     int viewId,
-    int contentSizeWidth,
-    int contentSizeHeight) {
+    float contentSizeWidth,
+    float contentSizeHeight) {
     super(viewId);
     mContentWidth = contentSizeWidth;
     mContentHeight = contentSizeHeight;
